update das ocorrências e correções de bugs

+++ script/ocorrencia.js
@@ -1,4 +1,5 @@
 const OcorrenciaForm = document.getElementById('ocorrencia-form');
+const CancelBtn = document.querySelector('.cancel-button');
 
 $(document).ready(function () {
 
@@ -7,8 +8,8 @@ $(document).ready(function () {
     var maskAntiga = 'AAA9999'; // Máscara para placas antigas
     var maskMercosul = 'AAA9A99'; // Máscara para placas Mercosul
 
-    var $placaVeiculo = $('#placa-veiculo');
-    var $placaCarreta = $('#placa-carreta');
+    var $placa = $('.placa');
+
 
     // Função para aplicar a máscara com base na condição
     function applyMask($element) {
@@ -21,25 +22,16 @@ $(document).ready(function () {
     }
 
     // Aplica a máscara inicial
-    $placaVeiculo.mask(maskAntiga);
-    $placaCarreta.mask(maskAntiga);
+    $placa.mask(maskAntiga);
 
     // Atualiza a máscara com base na entrada
-    $placaVeiculo.on('input', function () {
-        applyMask($placaVeiculo);
-    });
-
-    $placaCarreta.on('input', function () {
-        applyMask($placaCarreta);
+    $placa.on('input', function () {
+        applyMask($placa);
     });
 
     // Atualiza a máscara se o campo estiver vazio
-    $placaVeiculo.on('focus', function () {
-        applyMask($placaVeiculo);
-    });
-
-    $placaCarreta.on('focus', function () {
-        applyMask($placaCarreta);
+    $placa.on('focus', function () {
+        applyMask($placa);
     });
 });
 
@@ -99,7 +91,6 @@ document.addEventListener('DOMContentLoaded', function () {
     const inputs = document.querySelectorAll('.form-search input, .form-search select, .form-search textarea');
     inputs.forEach(input => {
         input.addEventListener('input', searchOcorrencias);
-        input.addEventListener('change', searchOcorrencias);
     });
 });
 
@@ -111,7 +102,6 @@ async function searchOcorrencias() {
     const descricao = document.getElementById('descricao-pesquisa').value;
     const status = document.getElementById('status-pesquisa').value;
     const dataDe = document.getElementById('data-de-pesquisa').value;
-    console.log(dataDe)
     const dataAte = document.getElementById('data-ate-pesquisa').value;
     const horaDe = document.getElementById('hora-de-pesquisa').value;
     const horaAte = document.getElementById('hora-ate-pesquisa').value;
@@ -147,17 +137,17 @@ function renderOcorrencias(ocorrencias) {
         const row = document.createElement('tr');
         row.setAttribute('data-id', ocorrencia.id_ocorrencia);
         row.innerHTML = `
-            <td>${ocorrencia.placa_veiculo}</td>
-            <td>${ocorrencia.placa_carreta}</td>
-            <td>${ocorrencia.cliente_nome}</td>
-            <td>${ocorrencia.motorista}</td>
-            <td>${ocorrencia.descricao}</td>
-            <td>${ocorrencia.status}</td>
-            <td>${ocorrencia.data_ocorrencia}</td>
-            <td>${ocorrencia.hora_ocorrencia}</td>
-            <td>${ocorrencia.usuario_login}</td>
+            <td class="placa-veiculo">${ocorrencia.placa_veiculo}</td>
+            <td class="placa-carreta">${ocorrencia.placa_carreta}</td>
+            <td class="cliente-nome" data-id="${ocorrencia.id_cliente}">${ocorrencia.cliente_nome}</td>
+            <td class="motorista">${ocorrencia.motorista}</td>
+            <td class="descricao">${ocorrencia.descricao}</td>
+            <td class="status">${ocorrencia.status}</td>
+            <td class="data-ocorrencia">${ocorrencia.data_ocorrencia}</td>
+            <td class="hora-ocorrencia">${ocorrencia.hora_ocorrencia}</td>
+             <td class="usuario-login" data-id="${ocorrencia.id_usuario}">${ocorrencia.usuario_login}</td>
             <td class="td-button">
-                <button type="button" class="edit-button">
+                <button type="button" class="edit-button" data-id="${ocorrencia.id_ocorrencia}">
                     <img width="30" height="30" src="/images/edit.png" alt="Editar">
                 </button>
                 <button type="button" class="delete-button" data-id="${ocorrencia.id_ocorrencia}">
@@ -168,10 +158,11 @@ function renderOcorrencias(ocorrencias) {
         tableBody.appendChild(row);
 
     });
-    addEventListenersToClientButtons();
+
 }
 
 async function updateOcorrenciaList() {
+    console.log("cheguei aqui!!!")
     try {
         const response = await fetch('/ocorrencia', { headers: { 'Accept': 'application/json' } });
         const data = await response.json();
@@ -212,17 +203,35 @@ async function deleteOcorrencia(ocorrenciaId) {
     }
 }
 
-function addEventListenersToClientButtons() {
+function addEventListenersToOcorrenciaButtons() {
+    const tableBody = document.getElementById('ocorrencias-table-body');
 
-    document.querySelectorAll('.delete-button').forEach(button => {
-        button.addEventListener('click', () => {
-            const ocorrenciaId = button.closest('tr').getAttribute('data-id');
+    tableBody.addEventListener('click', event => {
+        const target = event.target.closest('button');
+        if (!target) return; // Ignorar se o alvo não for um botão
+
+        const ocorrenciaId = target.getAttribute('data-id');
+
+        if (target.classList.contains('delete-button')) {
             deleteOcorrencia(ocorrenciaId);
-        });
+        } else if (target.classList.contains('edit-button')) {
+            editOcorrencia(ocorrenciaId);
+        }
     });
-
 }
 
+// Adiciona um event listener separado para o botão de salvar
+function addEventListenerToSaveButton() {
+    document.addEventListener('click', event => {
+        const target = event.target.closest('button.save-button');
+        if (!target) return; // Ignorar se o alvo não for um botão de salvar
+
+        const ocorrenciaId = target.getAttribute('data-id');
+        saveOcorrencia(ocorrenciaId);
+    });
+}
+
+addEventListenerToSaveButton();
 document.getElementById('export-pdf-button').addEventListener('click', function () {
     console.time("uzx");
 
@@ -255,4 +264,105 @@ document.getElementById('export-pdf-button').addEventListener('click', function 
     console.timeEnd("uzx");
 });
 
-addEventListenersToClientButtons();
+function editOcorrencia(ocorrenciaId) {
+    const listItem = document.querySelector(`tr[data-id="${ocorrenciaId}"]`);
+    document.getElementById('edit-placa-veiculo').value = listItem.querySelector('.placa-veiculo').innerText;
+    document.getElementById('edit-placa-carreta').value = listItem.querySelector('.placa-carreta').innerText;
+    //document.getElementById('edit-cliente-nome').value = listItem.querySelector('.cliente-nome').innerText;
+    document.getElementById('edit-motorista').value = listItem.querySelector('.motorista').innerText;
+    document.getElementById('edit-descricao').value = listItem.querySelector('.descricao').innerText;
+    document.getElementById('edit-status').value = listItem.querySelector('.status').innerText;
+
+
+    const dataOcorrencia = listItem.querySelector('.data-ocorrencia').innerText;
+    const formattedDataOcorrencia = formatDateForInput(dataOcorrencia);
+    document.getElementById('edit-data-ocorrencia').value = formattedDataOcorrencia;
+
+    // Preencher o campo de hora
+    document.getElementById('edit-hora-ocorrencia').value = listItem.querySelector('.hora-ocorrencia').innerText;
+
+    // Definir o data-id da div de edição
+    document.querySelector('.edit-section').dataset.id = ocorrenciaId;
+    document.querySelector('.save-button').dataset.id = ocorrenciaId;
+
+    // Preencher o select de usuário
+    const usuarioId = listItem.querySelector('.usuario-login').dataset.id;
+    console.log(usuarioId)
+    document.getElementById('edit-usuario-login').value = usuarioId;
+
+
+    // Supondo que você tem o `clientId` disponível
+    const clientId = listItem.querySelector('.cliente-nome').dataset.id;
+    console.log(clientId);
+
+    // Atualiza o Select2 com o valor do `clientId`
+    $('#edit-cliente-nome').val(clientId).trigger('change');
+
+    // Exibir a seção de edição
+    document.querySelector('.edit-section').style.display = 'flex';
+}
+
+
+
+if (CancelBtn) {
+    CancelBtn.addEventListener('click', () => {
+        document.querySelector('.edit-section').style.display = 'none';
+    });
+}
+
+
+function formatDateForInput(dateString) {
+    const [day, month, year] = dateString.split('/');
+    return `${year}-${month}-${day}`;
+}
+
+async function saveOcorrencia(ocorrenciaId) {
+
+    const placaVeiculo = document.getElementById('edit-placa-veiculo').value;
+    const placaCarreta = document.getElementById('edit-placa-carreta').value;
+    const clienteNome = document.getElementById('edit-cliente-nome').value;
+    const motorista = document.getElementById('edit-motorista').value;
+    const descricao = document.getElementById('edit-descricao').value;
+    const status = document.getElementById('edit-status').value;
+    const dataOcorrencia = document.getElementById('edit-data-ocorrencia').value;
+    const horaOcorrencia = document.getElementById('edit-hora-ocorrencia').value;
+    const usuarioId = document.getElementById('edit-usuario-login').value;
+
+    console.log(usuarioId)
+
+    //const ocorrenciaId = document.querySelector('#edit-section').dataset.idOcorrencia;
+
+    try {
+        const response = await fetch(`/update-ocorrencia/${ocorrenciaId}`, {
+            method: 'PUT',
+            headers: {
+                'Content-Type': 'application/json'
+            },
+            body: JSON.stringify({
+                'placa-veiculo-edit': placaVeiculo,
+                'placa-carreta-edit': placaCarreta,
+                'id-cliente-edit': clienteNome,
+                'motorista-edit': motorista,
+                'descricao-edit': descricao,
+                'status-edit': status,
+                'data-edit': dataOcorrencia,
+                'hora-edit': horaOcorrencia,
+                'id-usuario-edit': usuarioId
+            })
+        });
+
+        if (response.ok) {
+            alert('Ocorrência atualizada com sucesso.');
+            location.reload(); // Recarregar a página para ver as atualizações
+        } else {
+            const errorData = await response.json();
+            alert(`Erro: ${errorData.error}`);
+        }
+    } catch (error) {
+        console.error('Erro ao atualizar ocorrência:', error);
+        alert('Erro ao atualizar ocorrência.');
+    }
+
+}
+
+addEventListenersToOcorrenciaButtons();
